+++ source/neurons/cuba_lif.hpp
@@ -170,7 +170,7 @@ namespace hummus {
             if (type != spike_type::end_of_integration && potential >= threshold) {
                 // save spikes on final LIF layer before the Decision Layer for classification purposes if there's a decision-making layer
                 if (network->get_decision_making() && network->get_decision_parameters().layer_number == layer_id+1) {
-                    if (decision_queue.size() < network->get_decision_parameters().spike_history_size) {
+                    if (static_cast<int>(decision_queue.size()) < network->get_decision_parameters().spike_history_size) {
                         decision_queue.emplace_back(network->get_current_label());
                     } else {
                         decision_queue.pop_front();
@@ -321,7 +321,7 @@ namespace hummus {
                 
                 // save spikes on final LIF layer before the Decision Layer for classification purposes if there's a decision-making layer
                 if (network->get_decision_making() && network->get_decision_parameters().layer_number == layer_id+1) {
-                    if (decision_queue.size() < network->get_decision_parameters().spike_history_size) {
+                    if (static_cast<int>(decision_queue.size()) < network->get_decision_parameters().spike_history_size) {
                         decision_queue.emplace_back(network->get_current_label());
                     } else {
                         decision_queue.pop_front();
